tests/test_integration_compatibility.py
```python
import unittest
from src import integration_compatibility

class TestIntegrationCompatibility(unittest.TestCase):

    def setUp(self):
        self.excel_file = 'test_data.xlsx'
        self.email_platform = 'test_email_platform'

    def test_integrate_system(self):
        result = integration_compatibility.integrate_system(self.excel_file, self.email_platform)
        self.assertTrue(result)

    def test_excel_compatibility(self):
        result = integration_compatibility.check_excel_compatibility(self.excel_file)
        self.assertTrue(result)

    def test_email_platform_compatibility(self):
        result = integration_compatibility.check_email_platform_compatibility(self.email_platform)
        self.assertTrue(result)

if __name__ == '__main__':
    unittest.main()
```